Basic full text search capabilities (#62)
This is v1 of integrating full text search index into LanceDB.

The query API is roughly the same as before, except if the input is text
instead of a vector we assume that its fts search.

If `table` is a LanceDB LanceTable, then:

Build index: `table.create_fts_index("text")`

Query: `df = table.search("puppy").limit(10).select(["text"]).to_df()`

Here we use the tantivy-py package to build the index. We then use the
row id's as the full-text-search index's doc id then we just do a Take
operation to fetch the rows.

1. don't support incremental row appends yet. New data won't show up in
search
2. local filesystem only
3. requires building tantivy explicitly

---------

Co-authored-by: Chang She <chang@lancedb.com>